データベースの修正

+++ db/migrate/20181020044834_create_menu_drinks.rb
@@ -3,10 +3,10 @@ class CreateMenuDrinks < ActiveRecord::Migration[5.1]
     create_table :menu_drinks do |t|
       t.references :drink
       t.references :menu
-      t.integer :min_x
-      t.integer :min_y
-      t.integer :max_x
-      t.integer :max_y
+      t.decimal :min_x
+      t.decimal :min_y
+      t.decimal :max_x
+      t.decimal :max_y
       t.string :ocr_string
       t.string :ocr_language
